Separated history endpoints from notes
