+++ src/main.py
@@ -2,6 +2,7 @@ from fastapi import FastAPI
 
 from src.routers.user_router import router as user_router
 from src.routers.note_router import router as note_router
+from src.routers.history_router import router as history_router
 from src.routers.render_router import router as render_router
 from src.routers.folder_router import router as folder_router
 from src.routers.tag_router import router as tag_router
@@ -16,6 +17,10 @@ tags_metadata = [
     {
         "name": "Notes",
         "description": "Manage user's notes, CRUD operations, fix issues with notes, and go back to certain state of a note.",
+    },
+    {
+        "name": "History",
+        "description": "Manages notes' history and previous versions.",
     },
     {
         "name": "Folders",
@@ -52,9 +57,10 @@ app = FastAPI(
 
 app.include_router(user_router, prefix="/user", tags=["Users"])
 app.include_router(note_router, prefix="/note", tags=["Notes"])
+app.include_router(history_router, prefix="/history", tags=["History"])
 app.include_router(folder_router, prefix="/folder", tags=["Folders"])
 app.include_router(tag_router, prefix="/tag", tags=["Tags"])
-app.include_router(grammar_router, prefix="/tag", tags=["Grammar"])
+app.include_router(grammar_router, prefix="/grammar", tags=["Grammar"])
 app.include_router(render_router, prefix="/render", tags=["Render"])
 app.include_router(summarization_router, prefix="/ext", tags=["External"])
