+++ src/routers/note_router.py
@@ -15,7 +15,6 @@ from src.models.user import User
 from src.repositories.history_repository import HistoryRepository
 from src.repositories.note_repository import NoteRepository
 from src.schemas.folder_schema import ParentOut
-from src.schemas.history_schema import HistoryOut
 from src.schemas.note_schema import NoteOut, NoteIn, NoteUpdate
 from src.schemas.tag_schema import TagOut
 from src.services.history_service import HistoryService
@@ -37,8 +36,8 @@ router = APIRouter()
     status_code=status.HTTP_200_OK,
 )
 async def get_all_notes(
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
+        user: User = Depends(check_token),
+        session: AsyncSession = Depends(Connection.get_session),
 ):
     """
     This method is to get all notes in the database, where deleted field is set to be 0.
@@ -66,11 +65,11 @@ async def get_all_notes(
     status_code=status.HTTP_200_OK,
 )
 async def get_note_by_id(
-    note_id: int,
-    response: Response,
-    if_none_match: str | None = Header(default=None),
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
+        note_id: int,
+        response: Response,
+        if_none_match: str | None = Header(default=None),
+        user: User = Depends(check_token),
+        session: AsyncSession = Depends(Connection.get_session),
 ):
     """
     This method to get a note by its id.
@@ -106,9 +105,9 @@ async def get_note_by_id(
     status_code=status.HTTP_200_OK,
 )
 async def get_users_notes(
-    user_id: int,
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
+        user_id: int,
+        user: User = Depends(check_token),
+        session: AsyncSession = Depends(Connection.get_session),
 ):
     """
     This endpoint get all the user's notes available in the database with deleted field set to 0.
@@ -123,66 +122,6 @@ async def get_users_notes(
     return notes
 
 
-@router.get(
-    "/history/{note_id}",
-    summary="Get note's history  by its id",
-    description="This endpoint returns note's history if available inside the database",
-    response_model=list[HistoryOut],
-    response_description="The returned data is the history of a note",
-    responses={
-        200: {"description": "All note's history is returned successfully"},
-        404: {"description": "Note is not found"},
-    },
-    status_code=status.HTTP_200_OK,
-)
-async def get_note_history(
-    note_id: int,
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
-):
-    """
-    This endpoint to get the history and all the previous versions of a certain note.
-
-    :param note_id: The id of the note to get its history.
-    :param user: Check if the user is authorized to use the endpoint by checking the jwt token sent in the header.
-    :param session: This is the async session used to handle the database.
-    :return: The history of the note.
-    """
-    history_service = HistoryService(HistoryRepository(session))
-    versions = await history_service.get_note_versions(note_id)
-    return versions
-
-
-@router.get(
-    "/version/{version_id}",
-    summary="Get note's version",
-    description="This endpoint return a version of a note if available inside the database",
-    response_model=HistoryOut,
-    response_description="The returned data is a version of a note",
-    responses={
-        200: {"description": "The requested note's history returned successfully"},
-        404: {"description": "Note is not found"},
-    },
-    status_code=status.HTTP_200_OK,
-)
-async def get_note_old_version(
-    version_id: int,
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
-):
-    """
-    This endpoint returns a certain version of a note from its history.
-
-    :param version_id: The id of the version to be found.
-    :param user: Check if the user is authorized to use the endpoint by checking the jwt token sent in the header.
-    :param session: This is the async session used to handle the database.
-    :return: A certain version of a certain note.
-    """
-    history_service = HistoryService(HistoryRepository(session))
-    version = await history_service.get_version_by_id(version_id)
-    return version
-
-
 @router.post(
     "/",
     summary="Add new note",
@@ -194,9 +133,9 @@ async def get_note_old_version(
     status_code=status.HTTP_201_CREATED,
 )
 async def add_new_note(
-    note: NoteIn,
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
+        note: NoteIn,
+        user: User = Depends(check_token),
+        session: AsyncSession = Depends(Connection.get_session),
 ):
     """
     This method adds new note to the database.
@@ -232,10 +171,10 @@ async def add_new_note(
     status_code=status.HTTP_200_OK,
 )
 async def update_note(
-    note_id: int,
-    note: NoteUpdate,
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
+        note_id: int,
+        note: NoteUpdate,
+        user: User = Depends(check_token),
+        session: AsyncSession = Depends(Connection.get_session),
 ):
     """
     This endpoint to update available note's data, the note shall be available if not HTTPException 404 is raised.
@@ -276,9 +215,9 @@ async def update_note(
     status_code=status.HTTP_200_OK,
 )
 async def delete_note(
-    note_id: int,
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
+        note_id: int,
+        user: User = Depends(check_token),
+        session: AsyncSession = Depends(Connection.get_session),
 ):
     """
     This method to delete a note from the database if available, else it raises a 404 HTTPException.
